requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
alembic
passlib[bcrypt]
python-jose
